including eslint and a basic github workflow
