+++ .dockerignore
@@ -1,10 +1,15 @@
+.cache/
+.github/
 .vscode/
 node_modules/
 .dockerignore
+.editorconfig
 .env
+.eslintrc
 .git
+.gitattributes
 .gitignore
-.prettierignore
+.prettierrc
 Dockerfile
 LICENSE
 README.md
