+++ .prettierrc
@@ -1,0 +1,13 @@
+{
+  "overrides": [
+    {
+      "files":[
+        "**/*.js"
+      ],
+      "options": {
+        "singleQuote": true,
+        "trailingComma": "none"
+      }
+    }
+  ]
+}
