+++ .vscode/extensions.json
@@ -1,3 +1,3 @@
 {
-  "recommendations": ["esbenp.prettier-vscode", "ms-azuretools.vscode-docker"]
+  "recommendations": ["dbaeumer.vscode-eslint", "ms-azuretools.vscode-docker"]
 }
